+++ NEWS.md
@@ -9,6 +9,7 @@
 - Psi_Critic no longer used in Granier submodel (Stem vulnerability curve used instead)
 - New parameter Exp_Extract to model transpiration decrease in Granier submodel
 - Update growth/recruitment parameters
+- Removed fordyn dependency on input PET
 
 # Version 2.9.1
 - Allows species strings as input in forests objects